async refactor
