+++ unfetter-socket-server/src/vendor.d.ts
@@ -2,7 +2,6 @@ declare module NodeJS {
     interface Global {
         conn: any,
         unfetterconfigurations: any,
-        unfetteropenid: string,
-        unfettersocket: any
+        unfetteropenid: string
     }
 }
